refactor(taskFile): extract the source-location JSON locator
taskFile.ts (1397 LOC) bundled a bespoke ~220-LOC streaming JSON locator
(lineAndColumnAt, parseJsonString, findTasksArrayStart, extractTaskEntryLocations)
that maps tasks.json entries to line/column for diagnostics. These are pure
string functions with a single external consumer, so move them into
src/ralph/taskFileLocator.ts behaviour-unchanged and import the one public
function back. taskFile.ts shrinks by ~210 lines and the locator gains direct
unit tests (escape handling, nested-key rejection, brace/comma-in-string
robustness) that were previously only exercised transitively.
